+++ src/test/org/dbunit/dataset/datatype/DoubleDataTypeTest.java
@@ -88,7 +88,7 @@ public class DoubleDataTypeTest extends AbstractDataTypeTest
             new Double(0.666),
             new Double(5.49879),
             "-99.9",
-            new BigDecimal(1234),
+            new BigDecimal((double)1234),
         };
 
         Double[] expected = {
@@ -161,7 +161,7 @@ public class DoubleDataTypeTest extends AbstractDataTypeTest
             new Double(0.666),
             new Double(5.49879),
             "-99.9",
-            new BigDecimal(1234),
+            new BigDecimal((double)1234),
             "123",
         };
